+++ src/utils/api.js
@@ -48,6 +48,21 @@ export async function signUpUser({ name, email, password }) {
     });
 }
 
+export async function deleteUser({ token }) {
+  const requestOptions = {
+    method: "POST",
+    headers: {
+      Authorization: "Bearer " + token,
+    },
+  };
+
+  return fetch("https://gatovid.herokuapp.com/data/remove_user", requestOptions)
+    .then((data) => data.json())
+    .catch((error) => {
+      console.error("Error:", error);
+    });
+}
+
 export async function getUserData({ token }) {
   const requestOptions = {
     method: "POST",
